--- src/components/film-list.js
@@ -1,9 +1,0 @@
-
-export const createFilmListTemplate = () => {
-  return (
-    `<section class="films-list">
-    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-    <div class="films-list__container"></div>
-    </section>`
-  );
-};
