+++ src/components/menu.js
@@ -12,3 +12,13 @@ export const createMenuTemplate = () => {
   </nav>`
   );
 };
+
+export const createSortingTemplate = () => {
+  return (
+    `<ul class="sort">
+      <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
+      <li><a href="#" class="sort__button">Sort by date</a></li>
+      <li><a href="#" class="sort__button">Sort by rating</a></li>
+    </ul>`
+  );
+};
