--- src/components/show-button.js
@@ -1,6 +1,0 @@
-
-export const createShowButtonTemplate = () => {
-  return (
-    `<button class="films-list__show-more">Show more</button>`
-  );
-};
