+++ src/components/statistics.js
